06/05/2024 - First Part - Fix

+++ src/main/java/com/loparok/model/Basket.java
@@ -1,6 +1,6 @@
 package com.loparok.model;
 
-import java.sql.Date;
+import java.util.Date;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -17,7 +17,7 @@ public class Basket {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
+    private int id;
 
     @ManyToOne
     private Customer id_customer;
