+++ src/main/java/com/loparok/repository/BasketRepository.java
@@ -3,5 +3,5 @@ package com.loparok.repository;
 import com.loparok.model.Basket;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BasketRepository  extends JpaRepository<Basket, Long> {
+public interface BasketRepository  extends JpaRepository<Basket, Integer> {
 }
