+++ src/main/java/com/loparok/repository/Customer_have_addressRepository.java
@@ -3,5 +3,5 @@ package com.loparok.repository;
 import com.loparok.model.Customer_have_address;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface Customer_have_addressRepository  extends JpaRepository<Customer_have_address, Long> {
+public interface Customer_have_addressRepository  extends JpaRepository<Customer_have_address, Integer> {
 }
