+++ src/main/java/com/loparok/repository/Enterprise_have_procuctsRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.loparok.model.Enterprise_have_products;
 
-public interface Enterprise_have_procuctsRepository extends JpaRepository<Enterprise_have_products,Long>{
+public interface Enterprise_have_procuctsRepository extends JpaRepository<Enterprise_have_products, Integer>{
 
 }
