+++ src/main/java/com/loparok/repository/FeatureRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.loparok.model.Feature;
 
-public interface FeatureRepository extends JpaRepository<Feature, Long>{
+public interface FeatureRepository extends JpaRepository<Feature, Integer>{
 
 }
