+++ src/main/java/com/loparok/repository/Line_billRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.loparok.model.Line_bill;
 
-public interface Line_billRepository  extends JpaRepository<Line_bill, Long>{
+public interface Line_billRepository  extends JpaRepository<Line_bill, Integer>{
 
 }
