+++ src/main/java/com/loparok/repository/Product_have_discountRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.loparok.model.Product_have_discount;
 
-public interface Product_have_discountRepository  extends JpaRepository<Product_have_discount, Long>{
+public interface Product_have_discountRepository  extends JpaRepository<Product_have_discount, Integer>{
 
 }
